@@ -9,7 +9,7 @@ import java.util.Map;
 public class BetStratConstants {
 
     //Scrapping Service
-    public static final String SCRAPPER_SERVICE_URL = "http://47.237.107.158:8000/";
+    public static final String SCRAPPER_SERVICE_URL = "http://scrapper:8000/";
 
     public static final String SOCCERSTATS_BASE_URL = "http://www.soccerstats.com/";
     public static final String FCSTATS_BASE_URL = "http://www.fcstats.com/";
@@ -18,17 +18,17 @@ public class BetStratConstants {
     public static final String WORLDFOOTBALL_BASE_URL = "worldfootball.net";
     public static final String API_SPORTS_BASE_URL = "api-sports";
 
-    public static final String CURRENT_WINTER_SEASON = "2024-25";
-    public static final String CURRENT_SUMMER_SEASON = "2025";
+    public static final String CURRENT_WINTER_SEASON = "2025-26";
+    public static final String CURRENT_SUMMER_SEASON = "2026";
 
     public static final Integer DEFAULT_BAD_RUN_TO_NEW_SEQ = 5;
 
     public static final List<String> SEASONS_LIST = ImmutableList.of("2016","2016-17","2017","2017-18","2018","2018-19","2019","2019-20",
-            "2020","2020-21","2021","2021-22","2022","2022-23","2023-24");
+            "2020","2020-21","2021","2021-22","2022","2022-23","2023-24","2024-25");
 
     public static final List<String> SUMMER_SEASONS_LIST = ImmutableList.of("2016","2017","2018","2019","2020","2021","2022","2023","2024");
 
-    public static final List<String> WINTER_SEASONS_LIST = ImmutableList.of("2016-17","2017-18","2018-19","2019-20","2020-21","2021-22","2022-23","2023-24");
+    public static final List<String> WINTER_SEASONS_LIST = ImmutableList.of("2016-17","2017-18","2018-19","2019-20","2020-21","2021-22","2022-23","2023-24","2024-25");
 
     public static final List<String> SUMMER_SEASONS_BEGIN_MONTH_LIST = ImmutableList.of("January","February","March","April","May");
 
@@ -57,7 +57,7 @@ public class BetStratConstants {
     }};
 
     public static final List<String> LEAGUES_LIST = ImmutableList.of("https://www.worldfootball.net/schedule/chi-primera-division-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/per-primera-division-2025-apertura-spieltag/",
+        "https://www.worldfootball.net/schedule/per-primera-division-2025-clausura-spieltag/",
         "https://www.worldfootball.net/schedule/ecu-serie-a-2025-spieltag/",
         "https://www.worldfootball.net/schedule/est-meistriliiga-2025-spieltag/",
         "https://www.worldfootball.net/schedule/jpn-j1-league-2025-spieltag/",
@@ -67,34 +67,37 @@ public class BetStratConstants {
         "https://www.worldfootball.net/schedule/uru-primera-division-2025-apertura-spieltag/",
         "https://www.worldfootball.net/schedule/arg-copa-de-la-liga-profesional-2025-1-semestre-spieltag/",
         "https://www.worldfootball.net/schedule/col-primera-a-2025-apertura-spieltag/",
-        "https://www.worldfootball.net/schedule/aut-bundesliga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/bel-eerste-klasse-a-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/bul-parva-liga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/cro-1-hnl-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/cyp-first-division-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/cze-1-fotbalova-liga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/den-superliga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/eng-premier-league-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/fra-ligue-1-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/bundesliga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/gre-super-league-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/hun-nb-i-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/isr-ligat-haal-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/ita-serie-a-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/ned-eredivisie-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/pol-ekstraklasa-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/por-primeira-liga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/rou-liga-1-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/rus-premier-liga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/sco-premiership-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/srb-super-liga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/svk-super-liga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/svn-prvaliga-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/esp-primera-division-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/sui-super-league-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/tur-sueperlig-2024-2025-spieltag/",
-        "https://www.worldfootball.net/schedule/ukr-premyer-liga-2024-2025-spieltag/"
+        "https://www.worldfootball.net/schedule/bol-division-profesional-2025-spieltag",
+        "https://www.worldfootball.net/schedule/bra-serie-a-2025-spieltag/",
+        "https://www.worldfootball.net/schedule/nor-eliteserien-2025-spieltag/",
+        "https://www.worldfootball.net/schedule/swe-allsvenskan-2025-spieltag/",
+        "https://www.worldfootball.net/schedule/aut-bundesliga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/bel-pro-league-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/bul-parva-liga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/cro-1-hnl-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/cyp-first-division-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/cze-1-fotbalova-liga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/den-superliga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/eng-premier-league-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/fra-ligue-1-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/bundesliga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/gre-super-league-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/hun-nb-i-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/isr-ligat-haal-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/ita-serie-a-2025-2025-spieltag/",
+        "https://www.worldfootball.net/schedule/ned-eredivisie-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/pol-ekstraklasa-2026-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/por-primeira-liga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/rou-liga-1-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/rus-premier-liga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/sco-premiership-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/srb-super-liga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/svk-super-liga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/svn-prvaliga-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/esp-primera-division-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/sui-super-league-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/tur-sueperlig-2025-2026-spieltag/",
+        "https://www.worldfootball.net/schedule/ukr-premyer-liga-2025-2026-spieltag/"
         );
 
-    //missing: Bolivia Brasil Norway Sweden
 }
